Add files via upload
You are to implement a client-server application in Go for the management of tokens.
Token ADT
A token is an abstract data type, with the following properties: id, name, domain, and state. Tokens
are uniquely identified by their id, which is a string. The name of a token is another string. The
domain of a token consists of three uint64 integers: low, mid, and high. The state of a token
consists of two unit64 integers: a partial value and a final value, which is defined at the integer x
in the range [low, mid) and [low, high), respectively, that minimizes h(name, x) for a hash function
h. The hash h to use for this project is the SHA256 as below:
import (
"crypto/sha256"
"encoding/binary"
"fmt"
)
// Hash concatentates a message and a nonce and generates a hash value.
func Hash(name string, nonce uint64) uint64 {
hasher := sha256.New()
hasher.Write([]byte(fmt.Sprintf("%s %d", name, nonce)))
return binary.BigEndian.Uint64(hasher.Sum(nil))
}